Begin removing BlockEntityClientSerializable; add .sync helper
Apparently serialization has changed quite a bit for block entities:

https://bit.ly/3d1GLSy

The result of this is that they removed the BlockEntityClientSerializable class.

From the same issue, I found a snippet that should replace .sync and attached it as a method to GraveBlockEntity.

Next up is trying to rewrite the Nbt handlers..