Create playerCan* API; see description.
The code is quite verbose right now, checking whether an user has the
right permission level, whether grave robbing is enabled, if they own
the grave, if the grave even has an owner, different config options, etc.

These methods, which I've deemed the 'playerCan*' API, help alleviate
that.

playerCanAttemptRetrieve
  This method allows you to check whether the player meets the necessary
criteria to attempt to retrieve a grave. See HoneycombItemMixin and
ShovelItemMixin to see an example of how it's used.

playerCanOverride
  This method checks whether operator overriding is enabled, and if the
given player has the appropriate operator level to do so.

playerCanBreakGrave
playerCanUseGrave
  These two methods are similar -- they both check whether their
respective RetrievalTypes are active or not (along with if the user can
attempt to retrieve the grave).