Resolve onUse warning
Over at The Fabric Project I read that Minecraft doesn't use @deprecated
for truly deprecated code. In the context of a block, it means that the
method should be overriden and never called.

They also stated that there's a corresponding method inside
AbstractBlockState I should call instead.. but I was having issues with
that working.

Alas, both return ActionResult.PASS.. so I figured it couldn't hurt to
just do the same for now. I'll mess with it another time.

https://discord.com/channels/507304429255393322/507982478276034570/868218635473256448