Add configurable item decay (#26)
Configurable item decay allows the user to set a max decay percent
value. Every time a grave ages it will attempt to decay items by a
configurable amount.

The amount it decays is affected by how decayed the item already is;
the more decayed it is the quicker it'll decay.

[ Practical Example ]

Take for example a diamond sword with `61/1561 durability` (96.1% used
up) and a maxDecayPercent of `80`. The algorithm will use 96.1% of the
maxDecayPercent (76.8% in this case) to damage the item.

In this case, 61 (remaining durability) * .768 = 47 durability (rounded
up). This would leave the sword at 14/1561.

If the sword were at 1500/1561 durability it would damage it by 3.9% of
the maxDecayPercent (the sword would suffer 3.12% durability damage and
end up at 1453/1561).