Allow modfying multiple opts for client options
This allows you to add multipe options to the client options with one
command. For example:

> graves config set clientOptions add graves decayModifier

The above command would add both graves and deayModifier to the client
options. The 'remove' secondary option behaves in the same exact way.

I need to fix the suggestion provider for this but this should simplify
modifying the client options.