Add grave sinking mechanics
What does it mean to sink? Simply put, it'll iterate downwards until the
block below where the grave will spawn is not a sinkable block (based on
the configuration) and is within the world boundaries.

Graves can now sink in three different mediums:
- Air
- Water
- Lava
By default, graves will sink in Air and Water.