Fix client desync when failing to break block
In order to preserve the BlockEntity data of a grave that the client
isn't allowed to break (dictated by the server), the server config is
now sent to all clients upon joining. In addition, each client is
automatically subscribed to server config updates.

This allows the client to make informed decisions about whether it can
actually break the grave or not. And if it can't, the interaction will
be cancelled altogether. Thus, preserving the BlockEntity data.