Revert changes made in 25b21ca
I'm going to approach the problem differently:

I need to be able to retrieve a grave based on the GraveBlockEntity
itself, whether the grave exists in the world or not. I could modify the
existing `retrieve` method.. but instead, I'm going to make a `restore`
method as well.

The `retrieve` method will have some of its code split into methods that
will be shared between `retrieve` and `restore`. This should make the
code easier to read compared to the other solution I had came up with.