Leverage mstime during retrieveWithCommand @desc
If a new grave had spawned in the same location as an old one, and
someone tried to restore said old grave with the command, it would
simply remove the new grave and give you its contents.

Utilizing mstime (the time of creation) allows the code to discern
whether it's the same or a new grave in the given location; which
enables the code to fall back to creating a new grave entity instead.