Add switching drop types by holding shift
This adds a new option: shiftSwapsDropType. As the name hopefully
explains, holding shift while this setting is enabled and collecting
a grave will cause the drop type to be the opposite of what is
configured.

So, if your drop type is set to 'EQUIP' and you hold shift while
collecting the grave, it'll use the 'DROP' option and drop the items on
the ground.