(#92) Fix explosions not causing terrain damage
I had changed the code a bit whilst testing things and had forgotten to
change it back. Upon changing it back, I realized that the intended
implementation was sub-optimal anyways. 😂

This fix seems to be working as intended, though. Any block that is not a
grave will respond to the explosion as normal, and the graves themselves
are immune to explosions.