Make graves more resilient to non-player removal
Essentially, any grave that is not recovered via `onBreak` and `onUse`
will simply respawn when removed. This solves issues like explosions
destroying the grave.

This actually solves the same issue as ExplosionMixin.java but I'm
leaving the mixin there for redunancy in case there's some edgecase
where this code doesn't work and the mixin does.