Close #44; Add more realistic item decay
The decay now follows an S-curve. What this means is slower decay for
high durability items, decay speeding up the more damaged an item is,
and slowing down again toward the end of the item's durability.

The decay modifier still affects the speed of decay. A higher modifier
will speed up the decay process while a lower one will slow it down.