Fix potential server crash + grave spawning issue
This bug was somewhat niche. First, it requires a config option to have
been added to the Client Options array (options that are handled on a
per-client basis). For example, whether graves should spawn or not.

Next, a player will have to die while logging in.. before the mod has a
chance to add their config to the `clientConfigs` map. This would cause
the server to crash and stop the grave from ever spawning.