Add respectKeepInventory configuration option
Assuming KEEP_INVENOTRY is true...

If respectKeepInventory is false, a grave will spawn with items in
supported inventories. Any unsupported inventories will stay on the
player.

If respectKeepInventory is true, a grave will not spawn, and all items
will stay on the player.

And if KEEP_INVENTORY is false, a grave will spawn and unsupported
inventory items will drop (which is the current behavior).

This commit, and the last one, should resolve issue #114; though,
they'll have to enable KEEP_INVENTORY and set respectKeepInventory to
false.