(#115) Add better compatibility for Better Combat
There was an issue when Merge Order was set to `GRAVE` and the item in
the first hotbar slot of the grave was a two-handed weapon.

The issue resided in the fact that the `GRAVE` merge order runs two
`setInventory` calls. The initial one being the grave inventory, and the
second one being the player's inventory (at the time of retrieval).

The initial call would set the first hotbar slot as a two-handed weapon,
which appears to make the off-hand appear empty in the second call. And
that would cause the code to set the off-hand again (with air).

What this change does is two things:
1. The off-hand can't be set to AIR anymore.
2. The equipped off-hand is cached between `setInventory` calls.

Caching the off-hand between calls ensures the program accurately
understands what item is in the off-hand between calls, even if it
appears empty (due to the two-handed weapon).