Change publish system and updated things; see desc
I figured out what the issue was with publishing earlier: the loom
version. I guess GitHub was using a newer gradle version which didn't
work well with the old loom version.

But I've found out from this experience that I don't like having to make
a million commits to test out things with GitHub actions. So I'll be
opting for Modmuss50's Mod Publish Plugin from now on.

Anywho, I changed the stonecutter version back to 0.5.2, the fabric
loader to 0.16.10, gradle to 8.13, and loom to 1.10-SNAPSHOT.