Remove CommandContextData.java
This was largely used in the old config setter commands. The only
remaining use was that it was determining whether the reset and reload
commands were being issued to the server or client.

I moved that functionality to its own function and removed the class
entirely. It was janky to begin with and unnecessary now.