Remove respectKeepInventory option
The way keep inventory works with this mod is that all supported
inventories' items are stored in the grave while the rest remain on the
player. I added `respectKeepInventory` as a means to just disable grave
spawning when keep inventory is enabled... but the same can be
accomplished just by toggling the `graves` option anyway.